be/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
openai>=1.30
faiss-cpu>=1.8
numpy>=1.26
# 스크립트용 의존성 (이미 갖고 있으면 중복 OK)
pdfplumber>=0.10
pdfminer.six>=20221105
pdf2image>=1.17
pytesseract>=0.3.10
camelot-py[cv]>=0.11
opencv-python>=4.10
Pillow>=10.2
